# Conflicts:
#	CMakeLists.txt
#	server/CMakeLists.txt
#	server/internal/CMakeLists.txt
#	server/internal/entities/include/Task.hpp
#	server/internal/entities/include/User.hpp
#	server/internal/repository/CMakeLists.txt
#	server/internal/repository/include/SolutionRepository.hpp
#	server/internal/repository/include/TaskRepository.hpp
#	server/internal/repository/src/TaskRepository.cpp
#	server/internal/repository/src/UserRepository.cpp
#	server/internal/repository/virtual/ITaskRepository.hpp
#	server/internal/repository/virtual/IUserRepository.hpp
#	server/internal/service/CMakeLists.txt
#	server/internal/service/include/SolutionService.h
#	server/internal/service/include/TaskService.h
#	server/internal/service/include/UserService.h
#	server/internal/service/include/UserValidator.h
#	server/internal/service/src/SolutionService.cpp
#	server/internal/service/src/TaskService.cpp
#	server/internal/service/src/UserService.cpp
#	server/internal/service/src/UserValidator.cpp
#	server/internal/service/tests/CMakeLists.txt
#	server/internal/service/virtual/ISolutionService.h
#	server/internal/service/virtual/ITaskService.h
#	server/internal/service/virtual/IUserService.h

+++ server/internal/dbManager/include/dbManager.hpp
@@ -9,17 +9,23 @@
 #include <vector>
 #include <queue>
 #include <condition_variable>
+
 //#include "dotenv.h"
 //using namespace dotenv;
 class dbManager {
 public:
     dbManager();
+
     std::shared_ptr<pqxx::connection> connection();
-    void freeConnection(const std::shared_ptr<pqxx::connection>&);
+
+    void freeConnection(const std::shared_ptr<pqxx::connection> &);
+
 private:
-    const int POOL_SIZE = 10;
+    const size_t POOL_SIZE = 10;
     std::condition_variable m_condition;
+
     void createPool();
+
     std::queue<std::shared_ptr<pqxx::connection>> connection_pool;
     std::mutex m_mutex;
 };
